Invoices.Statistics module finished
